requirements.txt
Flask
#pyspark==3.5.3
#mysql-connector-python==9.4.0
##pysqlite3-binary
#delta-spark==3.3.2
#confluent-kafka[avro]==2.10.1
#avro==1.12.0
setuptools
ipywidgets
ipython-sql
pymysql
jupysql
duckdb-engine
grpcio
mcp
crewai
crewai[tools, litellm, openai] 
crewai-tools[mcp]
langsmith
langgraph
langchain
langchainhub
langchain_core
langchain_community
langchain-mcp-adapters
langchain-chroma 
langchain_openai
langchain_groq
langchain-huggingface
langchain-google-genai
langchain_ollama
chromadb
google-search-results 
numexpr 
text-generation
transformers 
sentence-transformers 
sentencepiece 
jinja2 
bitsandbytes 
accelerate
pypdf
pymupdf
faiss-cpu
python-dotenv
typesense
streamlit
duckduckgo-search
unstructured
ddgs
embedchain
openai-agents
docx2txt
tavily-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic[email]
jq
PyGithub